requirements.txt
numpy>=1.16.0
torch>=1.2.0
typing>=3.6.6
scikit_learn>=0.20.2
tensorboardX
torch-tb-profiler
lief
matplotlib
vt-py
python-dotenv